vector object, and function object
vector is an object holding the information of a field: it holds the
name of the field, and the data of the field as a "dimentioned
variable"(means it condisers the physical unit)
function object hold the name of the function, and the function itself,
when being constructed, an error function is being calculated and stored
in object properties. this object can calculate data fields.